refactor(rule-tester): refactor rule-tester (#33)
* chore: prepare v0.11.2

* chore: prepare v0.12.0

* refactor(rule-tester): refactor rule-tester
