--- tests/RuleTester/index.ts
@@ -1,4 +1,0 @@
-import { TSESLint } from "@typescript-eslint/experimental-utils";
-
-export const { RuleTester } = TSESLint;
-export * from "./resolveParser";
