+++ tests/lib/rules/use-name-space-import.spec.ts
@@ -1,9 +1,9 @@
-import { RuleTester, resolveParser } from "../../RuleTester";
+import { RuleTester } from "../../util";
 
 import rule from "../../../lib/rules/use-name-space-import";
 
 const tester = new RuleTester({
-  parser: resolveParser("@typescript-eslint/parser"),
+  parser: "@typescript-eslint/parser",
   parserOptions: {
     sourceType: "module",
     ecmaFeatures: {
